fixed more windows warnings
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@38228 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 31e1b97067e2b8bd94351e5bad695c6338050f07

+++ Rec/Tr/TrackCheckers/src/TrackCheckerNT.cpp
@@ -1251,8 +1251,8 @@ StatusCode TrackCheckerNT::fillMCGeneralInfo(
     // fill up to maxAssocTracks into our array
     for (int j = 0; (0 != track) && (j < maxAssocTracks);
 	track = revLink.next(), ++j) {
-      _assocTracks[nPart][j] = trackIdx[track];
-      _assocTrackWeights[nPart][j] = revLink.weight();
+      _assocTracks[nPart][j] = (float) trackIdx[track];
+      _assocTrackWeights[nPart][j] = (float) revLink.weight();
     }
   } // End loop over MCParticles
